refactoring (presentation layer)

+++ app/src/main/java/com/spinoza/cryptoapp/TimeUtils.kt
@@ -4,12 +4,12 @@ import java.sql.Timestamp
 import java.text.SimpleDateFormat
 import java.util.*
 
-fun convertTimeStampToTime(timeStamp: Int?): String {
+fun convertTimeStampToTime(timeStamp: Long?): String {
     val simpleDateFormat = SimpleDateFormat("HH:mm:ss", Locale.getDefault())
     simpleDateFormat.timeZone = TimeZone.getDefault()
     var result = ""
     timeStamp?.let {
-        result = simpleDateFormat.format(Date(Timestamp(timeStamp * 1000L).time))
+        result = simpleDateFormat.format(Date(Timestamp(timeStamp * 1000).time))
     }
     return result
 }
